requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pydantic==2.3.0
pydantic-core==2.6.3
pymongo==4.5.0
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.8.0
uvicorn==0.23.2
